requirements.txt
discord==1.0.1
spotipy==2.13.0
python-dotenv==0.14.0
